refactor(NavigationContext): renamed variables and functions
