+++ src/app/docs/t02-concepts/mod-01-levels-structure/mod-01-levels-structure.component.ts
@@ -5,7 +5,7 @@ import { AddHyphenPipe } from '../../../pipes/add-hyphen.pipe';
 
 // ▲ SERVICES ▲
 import { IntersectionService } from '../../../services/IntersectionObserver.service';
-import { LanguageService } from '../../../services/lenguaje.service';
+import { LanguageService } from '../../../services/navLanguage.service';
 import { Language } from '../../../services/language.types';
 
 // ▲ CONTENT ▲
@@ -41,12 +41,12 @@ export class Docs_T02_Mod01_Component implements OnInit, AfterViewInit {
   // ████ Fill Content (inner) ███
   ngOnInit(): void {
     // Initial content
-    this.write = content[this.currentLanguage].title_02.module_01;
+    this.write = content[this.currentLanguage].title_02.page_01;
 
-    // Suscribe to Language Service
-    this.languageService.language$.subscribe((language: string) => {
+    // Subscribe to Language Service
+    this.languageService.currentLanguage$.subscribe((language: string) => {
       if (this.isValidLanguage(language)) {
-        this.write = content[language].title_02.module_01;
+        this.write = content[language].title_02.page_01;
         this.currentLanguage = language;
       }
     });
