+++ src/app/layout/nav/nav.component.ts
@@ -3,7 +3,7 @@ import { Component, HostListener, AfterViewInit, OnInit } from '@angular/core';
 import { NavigationEnd, RouterLink, RouterLinkActive, RouterOutlet } from '@angular/router';
 import { ActivatedRoute, Router } from '@angular/router';
 import { IntersectionService } from '../../services/IntersectionObserver.service';
-import { TitleStateService } from '../../services/title.service';
+import { TitleService } from '../../services/navTitle.service';
 
 @Component({
   selector: 'app-nav',
@@ -23,7 +23,7 @@ export class NavComponent implements OnInit, AfterViewInit{
     private route: ActivatedRoute,
     private router: Router,
     private intersectionService: IntersectionService,
-    private titleStateService: TitleStateService
+    private TitleService: TitleService
   ) { }
 
   // Toogle
@@ -66,7 +66,7 @@ export class NavComponent implements OnInit, AfterViewInit{
       }
     });
     // Titles Toggle
-    this.titleStateService.currTitle$.subscribe(section => {
+    this.TitleService.currentTitle$.subscribe(section => {
       this.currTitle = section;
     })
   }
